requirements.txt
flask
gunicorn
requests
python-dotenv
anthropic
